fixing an error

+++ 9-binary_tree_height.c
@@ -7,8 +7,8 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int leftHeight;
-	int rightHeight;
+	size_t leftHeight;
+	size_t rightHeight;
 
 	leftHeight = binary_tree_height(tree->left);
 	rightHeight = binary_tree_height(tree->right);
